Fix 2 `User can't add child card in FireFox` #144

+++ src/js/plugins/inheritance.js
@@ -969,7 +969,7 @@
                   card.children.push(childCard);
                   childCard.parent = card;
                 }
-              } else {
+              } else if (!self.base.data.init) {
                 self.base.api.checklist.deleteItem(checklist.id, checkItem.id);
                 checklist.checkItems.splice(y, 1);
               }
